requirements.txt
python-dotenv>=1.0.0
starlette>=0.30.0
pydantic>=2
rich>=13
pytest>=8
numpy>=1.24.0
pandas>=2.0.0
setuptools>=68
requests>=2.31.0
bittensor>=7.0.0
bittensor-cli>=7.0.0
wandb>=0.18
pytz>=2025.2
scikit-learn>=1.3.0
PyYAML>=6.0.0
joblib>=1.3.0
tensorflow>=2.16.0
